test: arguments -> arguments + remote_reference
